requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
pdfplumber>=0.10.0
Pillow>=10.0.0
torch>=2.0.0
tqdm